+++ src/contexts/UserContext/actions.ts
@@ -1,8 +1,13 @@
-import onLogIn from '@axios/user/login';
-import onRegister from '@axios/user/register';
-import { LOGIN, LOGOUT, REGISTER, TOKEN } from '@utils/constantUser';
+import {
+  HEADERTOKEN,
+  LOGIN,
+  LOGOUT,
+  REGISTER,
+  TOKEN,
+} from '@utils/constantUser';
 import { removeStorage, setStorage } from '@utils/storage';
 import { Dispatch, useCallback } from 'react';
+import { onLogIn, onRegister, onLogOut } from '@axios/user';
 import { LoginUserInfo, RegisterUserInfo } from './types';
 
 const useUserProvider = (dispatch: Dispatch<any>) => {
@@ -13,16 +18,15 @@ const useUserProvider = (dispatch: Dispatch<any>) => {
       if (res.error.code) {
         throw new Error(`로그인 실패${res.error.code}`);
       }
-      // header에서 토큰이 어떻게 돌아오는지 확인해야됨
-      const xToken = 'x-auth-token';
+
       const header = await res.headers;
       const { id } = await res.data;
 
       dispatch({
         type: LOGIN,
-        user: { id, token: header[xToken] },
+        user: { id, token: header[HEADERTOKEN] },
       });
-      setStorage(TOKEN, header[xToken]);
+      setStorage(TOKEN, header[HEADERTOKEN]);
     },
     [dispatch]
   );
@@ -42,7 +46,8 @@ const useUserProvider = (dispatch: Dispatch<any>) => {
     [dispatch]
   );
 
-  const handleLogOut = useCallback(() => {
+  const handleLogOut = useCallback(async () => {
+    await onLogOut();
     dispatch({ type: LOGOUT });
     removeStorage(TOKEN);
   }, [dispatch]);
